% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSegProbs.R
\name{abbSegProb}
\alias{abbSegProb}
\title{Calculates probability of prespecified CN states being present}
\usage{
abbSegProb(
  gCnts,
  overDisp,
  segs = gCnts@metadata$segs,
  abbFrac = gCnts@metadata$segs$tumFrac,
  normFrac = 0.5,
  calcJointProb = TRUE,
  globalOnly = FALSE,
  ...
)
}
\arguments{
\item{gCnts}{Allele specific counts at individual loci in single cell data.  Produced by \code{\link{getAllelicExpression}}.}

\item{overDisp}{The beta-binomial over-dispersion parameter.  Usually calculated by \code{\link{calcOverDispersion}}.}

\item{segs}{GRanges object specifying segments with different copy number states.}

\item{abbFrac}{The major allele frequency expected for the abberant state in each segment.  See details.}

\item{normFrac}{The major allele frequency expected for the normal state in each segment.  See details.}

\item{calcJointProb}{Include a genome-wide estimate of state probabilities as well?}

\item{globalOnly}{Don't return segment specific calls, just the global model.}

\item{...}{Passed to \code{\link{calcStateProbs}}}
}
\value{
A GRanges object with each entry being a cell/CN segment combination with columns reporting the probability of the abbarent state (\code{probAbberant}).  If \code{globalOnly=TRUE} a data.frame is returned instead with probabilities for the abbarent state for each cell at a combined genome wide level.
}
\description{
Given a set of pre-specified CN states, such as copy number aberations in a tumour, calculates the posterior probability of each state, in each cell, for each segment.  This is a convenience function that calls \code{\link{calcStateProbs}} with parameters suited to identifying copy number changes in tumours.
}
\details{
The main thing to be specified is the GRanges object defining the copy number segments.  By default, the saved version specified by the user when running \code{\link{phaseSNPsFromCN}} is attempted to be used.  The model needs to know the expected major allele frequency for abberant and normal copy number states for each segment, which are assumed to be stored in the GRanges object with column names 'tumFrac' and 'normFrac'.  But they can also be specified by either giving a different column name, or explicitly specifying a value.  When values specified, they must be either a single value (if the same for all segments) or a vector of the same length as \code{segs}.

It is often the case that you are not interested in the individual segment probabilities for each cell, and just care about the joint probability of each cell having all the abbarent CN states.  The obvious example for this is the identification of tumour cells.  In this case, you can set \code{globalOnly=TRUE} and this function will return a table of global abbarent state probabilities for each cell.
}
