% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateByCluster.R
\name{aggregateByClusters}
\alias{aggregateByClusters}
\title{Collapse counts by clusters of cells}
\usage{
aggregateByClusters(
  gCnts,
  clusters = gCnts$clusterID,
  assays = c("A", "C", "G", "T", "Tot", "altCount", "refCount", "matCount", "patCount"),
  preserve = c("REF", "ALT", "geneID", "imprintingFDR", "imprinted", "errRate",
    "matASE", "matASE_postAlpha", "matASE_postBeta", "regionType", "altIsMum", "geneID",
    "geneName"),
  overlapBehaviour = c("all", "drop")
)
}
\arguments{
\item{gCnts}{Allele specific counts at individual loci in single cell data.  Produced by \code{\link{getAllelicExpression}}.}

\item{clusters}{List or named vector defining how cells are to be grouped.  If length matches \code{gCnts} assumed to directly indicate which clusters each row belongs to.  If a vector, names should be cellIDs and values cluster names.  If a list, list names should be cluster names.}

\item{assays}{Columns in \code{gCnts} to aggregate over.}

\item{preserve}{Columns in \code{gCnts} that should be copied across unchanged to the aggregated object.  Anything not present in gCnts in ignored.  E.g. geneID}

\item{overlapBehaviour}{When the same SNP overlaps multiple groupings, what should we do.  Options are 'all' or 'drop'.  See details.}
}
\value{
A GRanges object based on \code{gCnts}, with aggregated assays
}
\description{
Given a number of assays (e.g. counts) at a series of regions (e.g. SNPs), summarises assays by groupings of cells (e.g. clusters).
}
\details{
Clusters can be given as either a named list, or a named vector.  If it is a list, each list entry should contain cellIDs to be grouped.  When it is a vector, it should be a vector of cellIDs with cluster IDs as the names.

When a cell occurs in multiple multiple groupings there are multiple approaches to aggregation.  Current options (passed to \code{overlapBehaviour}) are 'drop' which ignores any cell that maps to more than one grouping, or 'all', which allocates the cell counts to all groupings that contain it.
}
\seealso{
aggregateByRegions aggregateByLists
}
