% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateByLists.R
\name{aggregateByLists}
\alias{aggregateByLists}
\title{Collapse counts by region and cluster}
\usage{
aggregateByLists(
  gCnts,
  assays,
  cellList = gCnts$cellID,
  regionList = gCnts$regionID,
  aggFun = sum
)
}
\arguments{
\item{gCnts}{Allele specific counts at individual loci in single cell data.  Produced by \code{\link{getAllelicExpression}}.}

\item{assays}{Columns in \code{gCnts} to aggregate over.}

\item{cellList}{Cell-like thing to aggregate over.  List or vector of same length as \code{gCnts}.}

\item{regionList}{Cell-like thing to aggregate over.  List or vector of same length as \code{gCnts}.}

\item{aggFun}{Function used for aggregation.}
}
\value{
A data.frame giving the combination of cellID and regionID from \code{cellList} and \code{regionList} and aggregated information in \code{assays}.
}
\description{
Given a count summarisation GRanges object with a cell like and region like element, aggregate values by vectors/lists of grouping factors.  This is typically something like cluster ID and/or genes, but this function works generally.
}
\details{
To aggregate on just cell-like or region-like variables, set the other entry to \code{1}.  That is, \code{cellList=1} and \code{regionList=gCnts$regionID} will aggregate by \code{regionID} ignoring any \code{cellIDs}.
}
