% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateByRegion.R
\name{aggregateByRegions}
\alias{aggregateByRegions}
\title{Collapse counts by regions}
\usage{
aggregateByRegions(gCnts, regions, assays, overlapBehaviour = c("all", "drop"))
}
\arguments{
\item{gCnts}{Allele specific counts at individual loci in single cell data.  Produced by \code{\link{getAllelicExpression}}.}

\item{regions}{GRanges or GRangesList object defining the regions in which to aggregate counts.}

\item{assays}{Columns in \code{gCnts} to aggregate over.}

\item{overlapBehaviour}{When the same SNP overlaps multiple groupings, what should we do.  Options are 'all' or 'drop'.  See details.}
}
\value{
A GRanges object based on \code{regions}, with aggregated counts for the supplied assays as columns.
}
\description{
Given a number of assays (e.g. counts) at a series of regions (e.g. SNPs), summarises assays by groupings of regions (e.g. genes).
}
\details{
Regions can be given as either a GRanges or GRangesList object.  If it is a GRangesList object, as created by \code{\link[GenomicFeatures]{exonsBy}} for example, the output will coerce this into a GRanges object using \code{range} (unless \code{returnMatricies} is TRUE).

When a SNP overlaps with multiple groupings there are multiple approaches to aggregation.  Current options (passed to \code{overlapBehaviour}) are 'drop' which ignores any SNP that maps to more than one grouping, or 'all', which allocates the SNPs counts to all groupings that overlap it.
}
\seealso{
aggregateByClusters aggregateByLists
}
