% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleleCounter.R
\name{alleleCounter}
\alias{alleleCounter}
\title{Runs allele Counter in parallel over a set of input and output files}
\usage{
alleleCounter(
  bams,
  refGenome,
  tgtLoci,
  outputs = NULL,
  f = 0,
  F = 0,
  x = TRUE,
  d = TRUE,
  m = 20,
  q = 200,
  bin = "alleleCounter",
  nParallel = 1,
  nChunks = 4,
  skipIfExists = TRUE
)
}
\arguments{
\item{bams}{The BAM files to get allele counts from.}

\item{refGenome}{The reference genome each BAM file was mapped using.}

\item{tgtLoci}{Locations to interogate as a GRanges object.  Either one per sample or if length 1 the same loci are used for all samples.}

\item{outputs}{The name of the file to write allele counts for each sample.  Lengths of outputs and bams must match.  If NULL, a temp file is used and deleted.}

\item{f}{Allelecounter -f param}

\item{F}{Allelecounter -F param}

\item{x}{Include the -x option?}

\item{d}{Include the -d option?}

\item{m}{Allelecounter -m param}

\item{q}{Allelecounter -q param}

\item{bin}{The location of the alleleCounter binary.}

\item{nParallel}{Number of processors to use.  Should be a multiple of number of samples.}

\item{nChunks}{When running in parallel, split into this many chunks per parallel thread.}

\item{skipIfExists}{If \code{outputs} already exists.  Load results from there instead of calculating from scratch.}
}
\value{
A list of GRanges objects, with each entry giving counts of A,C,G,T (and total) at the target loci provided.
}
\description{
Calculates the pileup using alleleCounter binary for a set of BAM files and target locations in parallel.  All parameters to alleleCounter can be either length 1 or the length of \code{bams}, in which case each parameter is matched to the corresponding BAM file.
}
\details{
If \code{outputs} is specified, the results will be saved to a file and reused instead of re-running alleleCounter if called again.

The default alleleCounter parameters are tuned for 10X output.
}
