% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateSNPs.R
\name{annotateSNPs}
\alias{annotateSNPs}
\title{Adds annotation to Variants}
\usage{
annotateSNPs(snps, gtf, autoChr = TRUE)
}
\arguments{
\item{snps}{GRanges for each variant.}

\item{gtf}{The GTF from which the annotation is to be loaded.}

\item{autoChr}{Try and automatically strip/add 'chr' to chromosome names to match gtf and snps.}
}
\value{
A GRanges
}
\description{
Given a gtf, annotates each variant (usually SNPS, but can be any one base genomic feature) based on what it overlaps.
}
