% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasGenes.R
\docType{data}
\name{biasedGenes}
\alias{biasedGenes}
\title{Genes with frequent allele specific bias}
\format{
A vector of gene symbols.
}
\usage{
data(biasedGenes)
}
\description{
This is a collection of genes that are either known to be imprinted in humans, or frequently have complex allele specific expression patterns that are not easily modelled.  Given this, the safest approach is to remove them when peforming inference on allele specific ratios
}
\keyword{datasets}
