% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCountMatricies.R
\name{buildCountMatricies}
\alias{buildCountMatricies}
\title{Constructs allele level count matricies}
\usage{
buildCountMatricies(
  counts,
  cellIDs = counts$cellID,
  regionIDs = counts$regionID,
  assays = c("matCount", "patCount")
)
}
\arguments{
\item{counts}{GRanges object summarising allele level expression.  Typically the output of \code{\link{getAllelicExpression}} or \code{\link{aggregateByRegions}}.}

\item{cellIDs}{For each row in \code{counts}, what cell does it belong to?  Must be of the same length as \code{counts}.}

\item{regionIDs}{For each row in \code{counts}, how should we label this region?  \code{as.character(counts)} is slow and often less useful, but should always work.}

\item{assays}{Which allele level expresion counts to convert to matricies.  Must exist as columns in \code{counts}.}
}
\value{
Region by cell sparse matricies containing counts for assays provided.  Returned as a list, unless only one assay given, in which case the matrix is returned.
}
\description{
Given allele level counts in single cell data, summarised at some genomic level (SNP, exon, gene), summarises counts into matricies.  These matricies have rows as genomic regions, columns as cell barcodes, and entries as the different allele level summaries specified in \code{assays}.
}
\details{
Any region that has no information in the single cell data is dropped.
}
