% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcASE.R
\name{calcASE}
\alias{calcASE}
\title{Calculates the allele specific expression of genes specified}
\usage{
calcASE(
  gCnts,
  normalCells,
  gns = NULL,
  priorKappa = NULL,
  overlapMax = TRUE,
  ...
)
}
\arguments{
\item{gCnts}{Allele specific counts at individual loci in single cell data.  Produced by \code{\link{getAllelicExpression}}.}

\item{normalCells}{Cells that you are very confident are normal.  If not specified, stored values used if present or all genes are set to have an expected allele fraction of 0.5}

\item{gns}{A GRanges object giving the coordinates of genes to use.  If NULL, load from annotation if \code{\link{annotateSNPs}} has been run.}

\item{priorKappa}{Value of kappa used for the beta prior.  If NULL, calculated using \code{\link{calcPriorASE}}.}

\item{overlapMax}{When two genes overlap a SNP, take the most extreme value.  If FALSE, takes the least extreme value.}

\item{...}{Passed to \code{\link{calcPriorASE}}.}
}
\value{
The input \code{gCnts} object, with an extra column giving an estimate of the maternal allele expression fraction, using posterior likelihood.  The alpha and beta values of the posterior ASE distribution are also added.
}
\description{
Determines what the default allelic balance of expression is for each gene, under the assumption of normality.  That is, given a set of cells that are very likely to be normal, the expression for each allele is aggregated across all cells and a posterior probability for the allelic ratio is calculated. Counts are modelled using a binomial distribution with beta prior, where the beta prior is pramaterised by it's mean (set to 0.5) and kappa = alpha + beta.  The shape of this posterior can be determined from the data, using \code{\link{calcPriorASE}}, although the default value of \code{priorKappa} should do a reasonable job.
}
\details{
For SNPs that don't overlap any of the genes provided in \code{gns}, a SNP level allele specific expression value is calculated using the same method.

Where a SNP is overlapped by more than one gene, the SNP is allocated the most extreme ASE value (i.e., biggest deviation from 0.5).  This behaviour can be changed by setting \code{overlapMax}.

If \code{gns} is NULL, gene information will be loaded from the annotation stored in \code{gCnts} from \code{\link{annotateSNPs}}.
}
