% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodModel.R
\name{calcEffectiveImbalance}
\alias{calcEffectiveImbalance}
\title{Calculate effective allelic imbalance}
\usage{
calcEffectiveImbalance(CNState, matASE, errRate)
}
\arguments{
\item{CNState}{The copy number state, expressed as a ratio of number of copies of the maternal allele out of the total number of copies.}

\item{matASE}{The allele specific expression bias, expressed as the number of counts expected from the maternal allele in a diploid cell.}

\item{errRate}{The error rate.}
}
\value{
The expected allelic imbalance for this set of parameters.
}
\description{
Given an error rate, copy number ratio, and allele specific expression, calculates the effective allelic imbalance that you would expect to observe in the data.
}
