% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcErrorRate.R
\name{calcErrorRate}
\alias{calcErrorRate}
\title{Determine error rate for groups of counts}
\usage{
calcErrorRate(gCnts, groupBy = gCnts$regionType)
}
\arguments{
\item{gCnts}{Allele specific counts at individual loci in single cell data.  Produced by \code{\link{getAllelicExpression}}.}

\item{groupBy}{How to group together things in \code{gCnts}.}
}
\value{
A vector giving the error rate by each group.
}
\description{
Assuming the genotype is correct, then only REF/ALT alleles should be observed and anything other than that is an error.  The number that are neither, sets a limit of the error rate.  This only measures sequencing errors, not getting the wrong allele due to ambient contamination, or mis-mapping.  There is also seldom sufficient power to detect error rates on a SNP by SNP basis.  So instead, SNPs are binned in some way (typically into intronic, exonic, and inter-genic) and the error rate for each group is calculated.
}
