% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodModel.R
\name{calcLikelihoods}
\alias{calcLikelihoods}
\title{Calcualtes the log-likelihood of various copy number states}
\usage{
calcLikelihoods(
  gCnts,
  overDisp,
  CNStates = c(pLoss = 1, mGain = 2/3, diploid = 1/2, pGain = 1/3, mLoss = 0),
  useASE = TRUE,
  marginaliseASE = FALSE,
  verbose = marginaliseASE,
  ...
)
}
\arguments{
\item{gCnts}{Allele specific counts at individual loci in single cell data.  Produced by \code{\link{getAllelicExpression}}.}

\item{overDisp}{The beta-binomial over-dispersion parameter.  Usually calculated by \code{\link{calcOverDispersion}}.}

\item{CNStates}{Specify which copy number states to test.  States are specified as ratios of number of maternal copies to total copies.  Names are used to name output columns.}

\item{useASE}{Should allele specific expression be incorporated into the model?}

\item{marginaliseASE}{Should we marginalise over the posterior distirbution for each gene's allele specific expression?}

\item{verbose}{Be verbose?}

\item{...}{Passed to \code{\link{likelihoodModel}}.}
}
\value{
\code{gCnts} with extra columns giving the negative log-likelihood for the model under each copy number state.  Extra columns named as \code{paste0('nLL_',names(CNStates))}.
}
\description{
Given allelic counts in individual cells (calculated by \code{getAllelicExpression}), determines the log-likelihood of each cell having a range of copy number configurations at each SNP.  By default, a range of the most plausible copy number states are calculated.  That is, maternal loss, paternal gain, diploid, maternal gain, and paternal loss.  However, any combination can be specified using the \code{CNStates} parameter.
}
\details{
The copy number states are given by the ratio of the number of maternal copies to total copies.  So maternal loss would be coded as \code{1/(0+1) = 1}, maternal gain \code{2/(1+2) = 0.33}, diploid \code{1/(1+1) = 0.5}, etc.  The names for \code{CNStates} are used to name output columns, so it is highly advisable to name them sensibly.  For instance, \code{CNStates=c(diploid=0.5,pLoH = 1)} will produce output columns \code{nLL_diploid} and \code{nLL_pLoH}.

The naiive model would be to compute a likelihood for each copy number state based on the ratio specified in \code{CNStates} using a binomial model.  This is insufficient for three reasons.  Firstly, sequencing contains errors.  This is most important in regions of loss of heterozygosity (\code{CNStates=0 or 1}), where without errors even a single read mapping to the lost allele would give probability zero.  The more accurately the error rate for each SNP can be specified, the more the likelihoods will represent reality.  Error rates are loaded from \code{gCnts} from the \code{errRate} column and set to \code{defaultErr} when not found.

Secondly, some genes have a persistent bias for one allele over another.  These genes will deviate from the 50/50 allele ratio of expression even in diploid regions.  When calculating the likelihood for the diploid state (\code{CNStates=0.5}), the allele specific expression of genes is taken into account.  This information is taken from the \code{matASE} column of \code{gCnts}, which gives the fraction of maternal counts expected in diploid cells for each SNP.  These values are usually calculated by running \code{\link{calcASE}} on \code{gCnts}.  This procedure will actually produce a probability distribution of likely allele specific expression values.  If \code{marginaliseASE} is set to \code{TRUE} the likelihood is obtained by marginalising over this distribution rather than using the most likely estimate.  Note that marginalising is slower and usually not worth the extra effort.  In the extreme case of imprinted genes, only one allele is expressed in all circumstances and so no information is left to determine the copy number state.  Therefore, imprinted genes are usually filtered out and allele specific expression is calculated by \code{\link{calcASE}}.  To prevent allele specific expression being considered at all, set \code{useASE=FALSE}.

The final reason the naiive model is insufficient is that transcription occurs in bursts on each allele and burst times are not perfectly correlated across alleles.  This means that more deviation from the ratio given in \code{CNStates} is expected that a binomial model would expect.  To account for this, we use a beta-binomial model, where the over-dispersion is parameterised by \code{overDispersion} and usually calcualted from the data using \code{\link{calcOverDispersion}}.
}
