% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMutationLikelihoods.R
\name{calcMutationLikelihoods}
\alias{calcMutationLikelihoods}
\title{Calculates the likelihood of each mutation in each cell}
\usage{
calcMutationLikelihoods(mCnts, overDisp, defaultErr = 0.01, mutFreq = 0.5)
}
\arguments{
\item{mCnts}{Somatic mutations GRanges.  Usually the output of \code{\link{loadCanPipeMuts}}.}

\item{overDisp}{The beta-binomial over-dispersion parameter.  Usually calculated by \code{\link{calcOverDispersion}}.}

\item{defaultErr}{Error rate to use when not specified by the 'errRate' column in \code{mCnts}.}

\item{mutFreq}{Frequency at which we expect to see a mutation in tumour cells.  Must be of length 1 or the same length as \code{mCnts}.}
}
\description{
Given a set of somatic mutations, calculates the likelihood of each cell containing a mutation under the null model (mutation not present + error rate) and the alternative model (mutation present at frequency specified).  The default is to assume that the mutation is on one of two copies (i.e., frequency 50%), but if the copy number configuration is known a more precise frequency can be given.  In general this tends not to matter as the posterior probability is based on the difference in log likelihood between the null and alt models.  If more mutant reads than expected by chance are present, that will be enough to favour the alt genotype even if the frequency is not correct.
}
