% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMutationPosteriors.R
\name{calcMutationPosteriors}
\alias{calcMutationPosteriors}
\title{Calculates posterior probability of each mutation}
\usage{
calcMutationPosteriors(mCnts, overDisp, defaultErr = 0.01, mutFreq = 0.5)
}
\arguments{
\item{mCnts}{Somatic mutations GRanges.  Usually the output of \code{\link{loadCanPipeMuts}}.}

\item{overDisp}{The beta-binomial over-dispersion parameter.  Usually calculated by \code{\link{calcOverDispersion}}.}

\item{defaultErr}{Error rate to use when not specified by the 'errRate' column in \code{mCnts}.}

\item{mutFreq}{Frequency at which we expect to see a mutation in tumour cells.  Must be of length 1 or the same length as \code{mCnts}.}
}
\value{
A table giving the posterior probability of the mutant genotype for each cell.
}
\description{
Given a set of mutations, calculates the posterior probability of each cell containing ALL mutations.  Likelihoods on which this is based are calculated using \code{\link{calcMutationLikelihoods}}, which will be run by this function if it has not been run previously.
}
