% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOverDispersion.R
\name{calcOverDispersion}
\alias{calcOverDispersion}
\title{Calculates burst kinetic over-dispersion}
\usage{
calcOverDispersion(gCnts, normalCells, marginaliseASE = FALSE, ...)
}
\arguments{
\item{gCnts}{Allele specific counts at individual loci in single cell data.  Produced by \code{\link{getAllelicExpression}}.}

\item{normalCells}{IDs of cells that can be considered normal.  If missing, use stored labelling of normal cells or consider all cells normal.}

\item{marginaliseASE}{Should we marginalise over the posterior distirbution for each gene's allele specific expression?}

\item{...}{Passed to \code{\link{likelihoodModel}}.}
}
\value{
The best fit for the over-dispersion.
}
\description{
The naive expectation is that without any copy number changes, expression should be equally distributed between alleles, with counts following a binomial distribution.  This basic expectation is incorrect for three reasons: errors, allele specific expression, and the burst like nature of transcription.  The model accounts for transcriptional bursts, where the tendency of each allele to have a burst of transcription not perfect correlated with the other allele, by using a beta-binomial model.  The extra dispersion introduced by this beta-binomial model recognises that even in CN neutral cells with no errors or allele specific bias in expression, we will still observe deviations from 50/50 reasonably often.
}
\details{
This function calibrates exactly how much extra dispersion to expect, by considering cells which are very likely to be normal and fitting a beta-binomial model.
}
\note{
The expectation is that \code{\link{calcASE}} has been run on \code{gCnts} before this function is called.  If it has not, \code{gCnts$matASE} must be defined (e.g. set to 0.5 for all) and \code{marginaliseASE} set to \code{FALSE}.
}
