% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPriorASE.R
\name{calcPriorASE}
\alias{calcPriorASE}
\title{Calculates the allele specific expression prior}
\usage{
calcPriorASE(gCnts, normalCells, gns = NULL, imprintCut = 0.01, expCut = 400)
}
\arguments{
\item{gCnts}{Allele specific counts at individual loci in single cell data.  Produced by \code{\link{getAllelicExpression}}.}

\item{normalCells}{IDs of cells that can be considered normal.  If missing, use stored labelling of normal cells or consider all cells normal.}

\item{gns}{A GRanges object giving the coordinates of genes to use.  If NULL, load from annotation if \code{\link{annotateSNPs}} has been run.}

\item{imprintCut}{Anything within this value of 0 or 1 are ignored.  Beta likelihood is not defined at 0 or 1.}

\item{expCut}{Consider genes with more than this many reads across all normal cells to be "highly expressed".}
}
\value{
The optimum value of kappa for the beta distribution prior.
}
\description{
The allele specific expression of each gene (or SNP) is estimated using a beta prior on a binomial likelihood.  As we don't expect any particular preference in allelic ratio, the mean of this beta prior should be 0.5.  This function sets the dispersion of this prior by fitting a beta distribution with mean 0.5 to the allelic ratios of high expression genes across cells highly likely to be normal.  That is, it looks at genes with extremely high expression where we can very confidently determine the allelic ratio and uses this to determine how much allele specific expression to expect for those genes where we don't have as much evidence.
}
\details{
We use the parameterisation of the beta distribution by mean mu and kappa = alpha + beta.  It is kappa that is optimised for and returned by this function.  If \code{\link{findImprinting}} has been run, genes marked as imprinted are not used.
}
\note{
Failure to properly specify a set of sensible normal cells (via \code{normalCells}) will lead to over-estimation of prior and reduced power to call copy number changes.
}
