% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStateProbs.R
\name{calcStateProbs}
\alias{calcStateProbs}
\title{Calculates posterior probability for a set of copy number states in genomic regions}
\usage{
calcStateProbs(
  gCnts,
  overDisp,
  regions = NULL,
  CNStates = c(pLoss = 1, mGain = 2/3, diploid = 1/2, pGain = 1/3, mLoss = 0),
  removeGroup = gCnts$geneID,
  removeN = 0,
  calcJointProb = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{gCnts}{Allele specific counts at individual loci in single cell data.  Produced by \code{\link{getAllelicExpression}}.}

\item{overDisp}{The beta-binomial over-dispersion parameter.  Usually calculated by \code{\link{calcOverDispersion}}.}

\item{regions}{GRanges object giving regions to calculate probabilities over.  Can also be a character vector specifying chromosomes.  If NULL, set to all chromosomes in \code{gCnts}.}

\item{CNStates}{What copy number states to test in each region.  Can be specified as a named vector (the default), in which case all regions are tested for the same CN states and the meaning is as in \code{\link{calcLikelihoods}}.  Alternatively, a pair of names can be given, specifying the names of columns in \code{regions} containing the allele fraction for tumour and normal respectively.}

\item{removeGroup}{Group things by this vector/list when removing.  Typically genes, but can be anything (position, whatever).  If NULL, no grouping, just remove SNPs.}

\item{removeN}{How many things to remove to check that we still have a significant result.}

\item{calcJointProb}{Include a genome-wide estimate of state probabilities as well?}

\item{verbose}{Be verbose?}

\item{...}{Passed to \code{\link{calcLikelihoods}}.}
}
\value{
A data.frame giving the posterior probabilities of each region being tumour and normal.  A genome wide posterior probability is also calculated and returned for each cell.
}
\description{
Given a set of regions, and a set of candidate copy number states to test, calculates the posterior probability of each copy number state, in each cell, in each region.
}
\details{
There are really two common use cases for this function.  In the first, the user has no prior expectation of what CN changes might be present in which cells.  The defaults are tuned to this scenario.  In this case, all plausibly detectable CN states are specified and posteriors calculated in each chromosome indepedently.

The other common use case is when a series of pre-defined copy number segments are supplied.  In this case each region has its own copy number state and \code{CNStates} should specify which columns of \code{regions} contain the expected BAF for the tumour and normal state.

By default, this will simply calculate the posterior probabilities with all available data.  However, if \code{removeN} is set to something greater than 0 a second calculation is performed with \code{removeN} bits of data removed per cell.  The logic here is that we want to know if the posterior probability is highly dependent on just a few regions.  If so, then it may be an artefact of the allele specific expression in those regions.  The default will try and remove the \code{removeN} genes that have the largest contribution to the posterior probability for each cell.  However, you can remove the top \code{removeN} SNPs or some other region by specifying a different grouping to \code{removeGroup}.
}
