% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeGenomeVersion.R
\name{changeGenomeVersion}
\alias{changeGenomeVersion}
\title{Convert positions between reference genomes}
\usage{
changeGenomeVersion(loci, chainPath)
}
\arguments{
\item{loci}{Locations to be lifted over.  Should be a GRanges object where each entry has width 1.}

\item{chainPath}{The path to the lift over chain to convert between coordinates.}
}
\value{
\code{loci} converted to new genome annotation.
}
\description{
Given a set of locations (SNPs, SNVs, etc) generated using one reference genome, uses the chain file located at \code{chainPath} to convert them to another reference system.
}
