% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Runs at startup}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library name.}

\item{pkgname}{Package name.}
}
\value{
Nothing.
}
\description{
Checks if singularity is installed on system and sets the binary paths appropriately.
}
