% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExAcSNPs.R
\docType{data}
\name{ExAcSNPs}
\alias{ExAcSNPs}
\title{ExAC SNPs hg19}
\format{
An object of class \code{"GRanges"}; see \code{\link[GenomicRanges]{GRanges}}.
}
\usage{
data(ExAcSNPs)
}
\description{
ExAC SNPs mapped to hg19.  These are taken from the HoneyBADGER data files (https://github.com/JEFworks/HoneyBADGER) which are pruned to only include exonic SNPs.
}
\keyword{datasets}
