% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCells.R
\name{filterCells}
\alias{filterCells}
\title{Filter cells}
\usage{
filterCells(
  phCnts,
  clusterIDs = NULL,
  passCellIDs = NULL,
  normIDs = NULL,
  regionsToKeep = c("Intronic", "Exonic"),
  genesToDrop = biasedGenes,
  dropImprinted = TRUE,
  dropInsane = TRUE,
  dropUninformative = TRUE,
  dropZeroCoverage = TRUE,
  segs = phCnts@metadata$segs,
  verbose = TRUE
)
}
\arguments{
\item{phCnts}{The counts to be filtered.  Usually produced by \code{\link{getAllelicExpression}}}

\item{clusterIDs}{Map of cluster IDs to cell IDs.  Either a vector or list, named by cluster ID, with entries being cell IDs.}

\item{passCellIDs}{The IDs of the cells to keep.  Derived from \code{clusterIDs} if given and this is set to \code{NULL}.}

\item{normIDs}{The IDs of the cells or clusters to mark as likely normal. If NULL, nothing stored.}

\item{regionsToKeep}{What types of regions should be kept?}

\item{genesToDrop}{Drop SNPs in these genes.  The defaults are genes that almost always have some hard to predict allele specific expression (either imprinting, or a bias) and are best ignored.}

\item{dropImprinted}{Drop imprinted regions.  Attempts to identify them using \code{\link{findImprinting}} if not already calculated.}

\item{dropInsane}{Drop SNPs that have failed sanity checks.}

\item{dropUninformative}{Drop SNPs marked as uninformative (i.e., couldn't be phased).}

\item{dropZeroCoverage}{Drop entries with no coverage of maternal or paternal allele.}

\item{segs}{CN segments to aggregate information for.  Only used if \code{verbose=2}.}

\item{verbose}{Report summary stats.}
}
\value{
A GRanges object with filtering done.
}
\description{
Counting of alleles using \code{\link{getAllelicExpression}} treats all barcodes as equal.  Usually we are not interested in most barcodes and only those that pass QC.  This function filters out useless barcodes, incorporates cluster information if available, and applies a few other sensible filters to produce a set of high quality counts.
}
\details{
If \code{clusterIDs} is provided, as either a list or named vector mapping cluster IDs to cell IDs, the permitted cell IDs will be derived from this object if not explicitly specified by \code{passCellIDs}.  The specification format is pretty flexible, so any kind of named vector or list that maps cellIDs to clusterIDs will probably work.

If \code{normIDs} are provided this information will be stored in the returned object and also used to calculated imprinted genes if needed and \code{dropImprinted=TRUE}.
}
