% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHetSNPs.R
\name{findHetSNPs}
\alias{findHetSNPs}
\title{Find heterozygous single nucleotide polymorphisms}
\usage{
findHetSNPs(
  bam,
  refGenome,
  outVCF = NULL,
  minCoverage = 10,
  errRate = 0.01,
  BAF_lim = c(0.2, 0.8),
  FDR = 0.05,
  minDeviation = 0.1,
  nParallel = 1,
  ...
)
}
\arguments{
\item{bam}{Path to BAM file to call variants from.}

\item{refGenome}{Path to reference genome used to map BAM.}

\item{outVCF}{The path of the VCF file containing SNPs to be created.}

\item{minCoverage}{Any location with coverage below this will not be reported.}

\item{errRate}{Assumed error rate for sanity check that location is heterozygous.  This check is disabled if \code{errRate=0}}

\item{BAF_lim}{Any SNP with BAF outside this range will be discarded.}

\item{FDR}{False discover rate to accept.}

\item{minDeviation}{Minimum amount that the BAF must deviate from 0.5 before we reject it.}

\item{nParallel}{How many threads to use.}

\item{...}{Other things passed to \code{\link{findVariants}}}
}
\value{
A GRanges object containing heterozygous SNPs.
}
\description{
Uses bcftools to find heterozygous SNPs.  See \code{\link{findVariants}}.
}
\details{
All variants with reasonable coverage (\code{minCoverage}) and BAF consistent with a het SNP (\code{BAF_lim}) are identified.  These are then further filtered and any with a significant deviation from 0.5 BAF are removed.  Significant deviation is defined by binomial FDR less than \code{FDR} and 0.5 - \code{minDeviation} < BAF < 0.5 + \code{minDeviation}.

The defaults assume that \code{bam} comes from normal tissue.  If instead this comes from tumour tissue (or some other type of mixed genotype tissue), you should change some parameters.  The difference with calling SNPs from tumour sequencing SNPs will not have BAF of 0.5, so the \code{minDeviation} filter is not appropriate and should be set to \code{minDeviation = 0.5}.  You may also want to adjust `BAF_lim` to be more permissive and/or adjust \code{errRate}.
}
