% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findImprinting.R
\name{findImprinting}
\alias{findImprinting}
\title{Find imprinted regions}
\usage{
findImprinting(
  gCnts,
  normalCells,
  gns = NULL,
  useGenes = TRUE,
  nullDiff = 0.3,
  FDR = 0.05
)
}
\arguments{
\item{gCnts}{Allele specific counts at individual loci in single cell data.  Produced by \code{\link{getAllelicExpression}}.}

\item{normalCells}{IDs of cells that can be considered normal.  If missing, use stored labelling of normal cells or consider all cells normal.}

\item{gns}{A GRanges object giving the coordinates of genes to use.  If NULL, load from annotation if \code{\link{annotateSNPs}} has been run.}

\item{useGenes}{Should we use genes?  If FALSE will consider each SNP independently and ignore gene information.}

\item{nullDiff}{The deviation from 0.5 to use for the maternal allele expression fraction under the null hypothesis.  E.g. 0.3 implies a null of 0.2 or 0.8.  See details.}

\item{FDR}{False discover rate to accept.}
}
\value{
A modified version of \code{gCnts} with columns indicating imprinting.  A table with details underlying the calls is also added to \code{gCnts@metadata$imprinted}.
}
\description{
Given allele specific expression at het SNPs and gene definitions, identifies genes or individual SNPs that have only one allele expressed across all normal cells.  These regions could either be due to imprinting or mis-called heterozygous SNPs (see note).
}
\details{
Imprinted genes are determined using a binomial test, against the null of strong allele specific expression (but not imprinting).  The strength of the null hypothesis is determined by \code{nullDiff}.  This sets how far from 0.5 the expected allele frequency should be for the null distribution.  For a gene that appears to be imprinted with no maternal expression, the null tests that the maternal fraction  is less than \code{0.5 - nullDif}.  For a gene with no paternal expression, it would test \code{0.5+nullDiff}.

If genes are specified by \code{gns} then an extra column, geneImprinted, will be added to the output indicating if there is evidence of imprinting at the gene level.
}
\note{
A gene appearing in this list is either evidence of imprinting **or** evidence that the heterozygous SNP in this region is not actually heterozygous.  For this reason, genes with higher values of \code{nSupportingSNPs} are more reliable.  If all you want to do is call CN segments, it doesn't matter, just remove everything that passes.  If you are interested in the biology, look closely at the SNPs underlying the call and treat those with only 1 or 2 SNPs with suspicion.  Even when many SNPs support a gene, make sure that the call is not being driven by high coverage of one SNP with a trivial contribution from the others.

This differs from the results of \code{\link{calcASE}} in that the allele specific expression prior is calculated excluding imprinted genes.  The ideal workflow is to first run \code{findImprinting} and pass the resulting imprinted genes to \code{\link{calcASE}}.
}
