% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVariants.R
\name{findVariants}
\alias{findVariants}
\title{Use BCFtools to find variants in a BAM file}
\usage{
findVariants(
  bam,
  refGenome,
  minCoverage,
  BAF_lim,
  outVCF = NULL,
  chrsToProcess = c("X", 1:22),
  minMapQual = 35,
  minBaseQual = 20,
  minVarQual = 225,
  bin = "bcftools",
  skipIfExists = TRUE,
  nParallel = 1,
  nChunks = 1,
  ...
)
}
\arguments{
\item{bam}{Path to BAM file to call variants from.}

\item{refGenome}{Path to reference genome used to map BAM.}

\item{minCoverage}{Any location with coverage below this will not be reported.}

\item{BAF_lim}{Any SNP with BAF outside this range will be discarded.  Vector of length tow, e.g. \code{BAF_lim=c(0.2,0.8)}.}

\item{outVCF}{The path of the VCF file containing SNPs to be created.}

\item{chrsToProcess}{Which chromosomes to process?  If NULL, all chromosomes in the BAM header are used.  Any chromosomes not present in the BAM header will be dropped.  If none found, tries prepending 'chr' to this before failing.}

\item{minMapQual}{Minimum map quality to pass to -q flag of bcftools mpileup.}

\item{minBaseQual}{Minimum base quality to pass to -Q flag of bcftools mpileup.}

\item{minVarQual}{Filter variants with quality scores below this value.}

\item{bin}{Path to BCF tools binary.}

\item{skipIfExists}{Skip running the SNP caller if the VCF already exists.}

\item{nParallel}{How many threads to use.  Note that the final variants called depends (very mildly) on the genomic chunks processed together.  As such, it is not advisable to set this much larger than the number of chromosomes.}

\item{nChunks}{When running in parallel, split into this many chunks per parallel thread.}

\item{...}{Absorb unused parameters.}
}
\value{
A GRanges object containing all variants matching specified parameters.
}
\description{
Find any base with decent coverage and BAF within a range and save them a VCF. If an output filename is given via \code{outVCF}, an earlier result will be loaded if it exists.
}
\details{
This is mostly just a wrapper around external calls to bcftools, which obviously need to be installed for this to work.
}
