% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCoverageAndBAF.R
\name{generateCoverageAndBAF}
\alias{generateCoverageAndBAF}
\title{Generate coverage (logR) and BAF from BAM file}
\usage{
generateCoverageAndBAF(
  BAM,
  refGenome,
  hSNPs = SNPs1k,
  AF = hSNPs$AF,
  outPath = NULL,
  gcFile = NULL,
  alleleCounterParams = list(f = 3, F = 3852, m = 20, q = 35),
  chrs = c(1:22, "X"),
  errRate = 0.01,
  abbCellFrac = 0.95,
  minHomProb = 0.05,
  minHetAF = 0.1,
  runASCAT = FALSE,
  nParallel = 1
)
}
\arguments{
\item{BAM}{The BAM file to calculate coverage and BAF on.}

\item{refGenome}{Path to reference genome used to map BAM.}

\item{hSNPs}{The heterozygous SNPs to investigate.  Defaults to 1000 genomes SNPs.  Must be a GRanges object with REF and ALT columns.}

\item{AF}{The population allele frequency of the alternate allele.  If NULL, defaults to 0.05}

\item{outPath}{File name used to storing genotyping calculations.  If left at NULL, results are not saved and cannot be reloaded in subsequent runs.}

\item{gcFile}{ASCAT GC correction file for the SNPs provided.  If given, used to GC correct logR.}

\item{alleleCounterParams}{Parameters to be used when running allele counter.}

\item{chrs}{Which chromosomes to plot and/or run ASCAT on.}

\item{errRate}{Assumed sequencing error rate.}

\item{abbCellFrac}{Assumed fraction of abberant cells for determining heterozygous SNPs.  Higher values are more stringent.}

\item{minHomProb}{Any SNP with posterior probability of homozygosity in the germline greater than this will be code as homozygous.}

\item{minHetAF}{Any SNP with BAF below this will be coded as homozygous.}

\item{runASCAT}{Should we run ASCAT?}

\item{nParallel}{How many threads to use.}
}
\value{
The hSNPs object with extra columns, particularly total coverage and BAF.  If ASCAT has been run, columns associated with the fit are also added and the full ASCAT output is stored at \code{hSNPs@metadata$ascat.output}.
}
\description{
The most "raw" way to look at copy number status of a BAM file is to calculate the coverage and alternate allele frequency at heterozygous SNPs.  This function uses alleleCounter to generate this from a BAM file, for a set of heterozygous SNPs.  You can specify the SNPs however you like, but usually this would either be using a pre-specified panel of SNPs, or using the \code{\link{findHetSNPs}} function to identify potential heterozygous SNPs directly.
}
\details{
Will optionally run ASCAT to infer CN segments (in unmatched mode) and produces a basic plot to visualise the raw data.  If ASCAT is run, there is some extra information that is needed.  These values almost all relate to guessing if a given SNP is heterozygous in the germline.  The allele frequency \code{AF} is used as a prior and posterior probabilities of heterozygosity and homozygosity (in the germline) is determined using the provided error rate (for homozygosity) and maximal abbarent cell fraction (\code{abbCellFrac} for heterozygosity).  Any SNP with posterior probability of homozygosity greater than \code{minHomProb} or observed BAF less than \code{minHetAF} will be coded as homozygous.  This errs on the side of marking things as homozygous if the data is not clear, which is the correct thing to do.
}
