% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllelicExpression.R
\name{getAllelicExpression}
\alias{getAllelicExpression}
\title{Get allele specific expression from RNA BAM files at base pair resolution}
\usage{
getAllelicExpression(
  loci,
  refGenome,
  bams,
  labels = names(bams),
  outputs = NULL,
  minCounts = 0,
  assayType = c("10X", "SS2", "Other"),
  matAlleleCol = "matAllele",
  patAlleleCol = "patAllele",
  segs = loci@metadata$segs,
  errRate = 0.02,
  verbose = TRUE,
  nParallel = 1,
  ...
)
}
\arguments{
\item{loci}{GRanges object containing heterozygous SNP locations to measure.  Must have REF and ALT.}

\item{refGenome}{Reference genome used to align BAM files.}

\item{bams}{Aligned BAM files}

\item{labels}{Unique label for each BAM.  Name of \code{bams} used if available.  If not, generated automatically.}

\item{outputs}{File names to store output for each BAM file in. See \code{\link{alleleCounter}}.}

\item{minCounts}{Any SNP that doesn't have at least this many counts from both alleles across all cells is marked as \code{passSanity=FALSE}.}

\item{assayType}{The type of expression assay run.  Used to determine the default parameters passed to \code{\link{alleleCounter}}.  These can always be over-riden by directly specifying allelelCounter params yourself.}

\item{matAlleleCol}{Column in \code{loci} indicating which base is maternally derived.}

\item{patAlleleCol}{Column in \code{loci} indicating which base is maternally derived.}

\item{segs}{CN segments.  Used for summary stats if present and \code{verbose=2}.}

\item{errRate}{Assumed sequencing error rate.  Only used for summary stats.}

\item{verbose}{Be verbose?  0/FALSE reports nothing, 1/TRUE reports a minimal summary, 2 for a detailed summary.}

\item{nParallel}{Number of threads to use.}

\item{...}{Passed to alleleCounter.}
}
\value{
The input GRanges object \code{loci}, but expanded to have one line per cell in the expression data and with columns added to give the number of counts in that cell, at that location in the expression data.
}
\description{
Given a set of individual nucleotide genomic coordinates, interrogates all locations in a set of RNA bam files and generates a cell specific allele count.  The output is \code{loci}, with extra columns added recording the number of counts for A,C,G,T, the refence/alternative alleles (as encoded by REF/ALT) and the maternal/paternal allele (if known).  The labels for each are determined by the \code{labels} param and derived from the BAM file name by default.  \code{loci} must contain columns named \code{REF} and \code{ALT} specifying the REF/ALT allele.
}
\details{
Each entry is given a regionID, which is given in the format chr:pos_refAllele/altAllele with the maternal allele (where known) coded as a lower case letter.
}
