% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCellBarcodes10X}
\alias{getCellBarcodes10X}
\title{Get filtered cellIDs from 10X data}
\usage{
getCellBarcodes10X(bams, labels = names(bams))
}
\arguments{
\item{bams}{The 10X BAM files.}

\item{labels}{The label for each 10X bam file.}
}
\value{
The barcodes that pass 10X filters.
}
\description{
Given BAM files and labels, as passed to \code{\link{getAllelicExpression}}, assumes the BAM files sit in a cellranger output folder and loads those barcodes that are in the filtered list.
}
