% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getColours}
\alias{getColours}
\title{Get colours for a factor}
\usage{
getColours(x)
}
\arguments{
\item{x}{factor or thing to be turned into factor.}
}
\value{
Vector of colours.
}
\description{
Given a factor, or something that can be made a factor, gets a vector of colours.  Named by unique values.
}
