% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getColoursN}
\alias{getColoursN}
\title{Get N distinct colours}
\usage{
getColoursN(n)
}
\arguments{
\item{n}{Number of colours}
}
\description{
Gets N distinct colours.  Taken from here https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}
