% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getErrorRate.R
\name{getErrorRate}
\alias{getErrorRate}
\title{Get error rate}
\usage{
getErrorRate(gCnts, defaultErr = 0.01)
}
\arguments{
\item{gCnts}{Allele specific counts at individual loci in single cell data.  Produced by \code{\link{getAllelicExpression}}.}

\item{defaultErr}{The error rate to use if none can be found.}
}
\value{
A vector of error rates.
}
\description{
Retrieves error rate previously calculated using \code{calcErrorRate} and falls back on a default if not present.
}
