% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRowColNums}
\alias{getRowColNums}
\title{Get number of cols/roms for plot grid}
\usage{
getRowColNums(nPlots, r = 4/3, nCol = NULL, nRow = NULL)
}
\arguments{
\item{nPlots}{Number of plots}

\item{r}{Aspect ratio.}

\item{nCol}{Number of columns.  Worked out if not fixed.}

\item{nRow}{Number of rows  Worked out if not fixed.}
}
\description{
Given a desired aspect ratio and a number of plots, decides on how many rows/columns are needed.
}
