% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getShapes}
\alias{getShapes}
\title{Get shape for a factor}
\usage{
getShapes(x)
}
\arguments{
\item{x}{factor or thing to be turned into factor.}
}
\value{
Vector of shapes.
}
\description{
Given a factor, or something that can be made a factor, gets a vector of shapes.  Named by unique values.  Recycled if more than 25.
}
