% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodModel.R
\name{likelihoodModel}
\alias{likelihoodModel}
\title{Likelihood model}
\usage{
likelihoodModel(
  obsMat,
  obsPat,
  CNState,
  overDisp,
  errRate,
  matASE,
  alpha,
  beta,
  defaultErr = 0.01,
  marginaliseASE = FALSE,
  minVarASE = 1e-04,
  verbose = marginaliseASE,
  ...
)
}
\arguments{
\item{obsMat}{Vector of observed maternal counts.}

\item{obsPat}{Vector of observed paternal counts.}

\item{CNState}{The copy number state, expressed as a ratio of number of copies of the maternal allele out of the total number of copies.}

\item{overDisp}{Over-dispersion parameter for beta-binomial distribution.}

\item{errRate}{The error rate.}

\item{matASE}{The allele specific expression bias, expressed as the number of counts expected from the maternal allele in a diploid cell.}

\item{alpha}{Alpha parameter of the prior on the allele specific expression ratio (\code{matASE}).}

\item{beta}{Beta parameter of the prior on the allele specific expression ratio (\code{matASE}).}

\item{defaultErr}{Default error rate to use if none given.}

\item{marginaliseASE}{Marginalise or not?}

\item{minVarASE}{If marginalising and the variance in the ASE is less than this, just use the central estimate.}

\item{verbose}{Show progress?}

\item{...}{Extra things passed to \code{integrate} if marginalising ASE.}
}
\value{
The negative log-likelihood for each observation.
}
\description{
Calcualtes the likelihood model, with or without marginalisation, ASE, etc.  For user facing wrapper see \code{\link{calcLikelihoods}}.
}
