% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCanpipeMuts.R
\name{loadCanPipeMuts}
\alias{loadCanPipeMuts}
\title{Load Mutations from CanPipe}
\usage{
loadCanPipeMuts(cavemanOutput)
}
\arguments{
\item{cavemanOutput}{Path to caveman mutation output file (the one ending flag.vcf.gz).}
}
\value{
GRanges file with mutations.
}
\description{
Loads somatic point mutations from Caveman output file (flag.vcf.gz) and format them as needed.  Only point mutations with FILTER=='PASS' are kept.
}
