% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCanpipeSNPs.R
\name{loadCanPipeSNP}
\alias{loadCanPipeSNP}
\title{Load SNPs from CanPipe}
\usage{
loadCanPipeSNP(cavemanOutput)
}
\arguments{
\item{cavemanOutput}{Path to caveman SNP output.}
}
\value{
GRanges SNP file.
}
\description{
Loads Caveman SNPs from canpipe file and format them as needed.
}
