% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchBAMs.R
\name{matchBAMs}
\alias{matchBAMs}
\title{Checks if BAM files have matching genotype.}
\usage{
matchBAMs(
  BAMs,
  refGenomes,
  snps = ExAcSNPs,
  outputs = NULL,
  liftOvers = NULL,
  is10X = FALSE,
  usesChr = FALSE,
  nParallel = 1,
  doPlot = TRUE,
  colPal = "Greens",
  ...
)
}
\arguments{
\item{BAMs}{The BAM files to genotype.  Names are used as labels.}

\item{refGenomes}{The FASTA file for the genome of each sample.}

\item{snps}{A GRanges object with SNPs to inspect.  Must contain REF and ALT columns.}

\item{outputs}{Store output SNP counts to prevent recalculation.  Vector of length to match \code{BAMs}.}

\item{liftOvers}{If one or more of the BAMs uses a different coordinate system to the reference, lift the coordinates over with these files.  Vector of length BAMs, with NAs for no-liftover.}

\item{is10X}{If a BAM is 10X, set to TRUE so it can be counted properly.}

\item{usesChr}{Do any samples have a 'chr' prefix on chromosome names.}

\item{nParallel}{Number of processors to use.  Should be a multiple of number of samples.}

\item{doPlot}{Make a plot of the results?}

\item{colPal}{Colour scheme to use.  Passed to \code{\link[RColorBrewer]{brewer.pal}}.}

\item{...}{Pass to \code{\link{alleleCounter}}.}
}
\value{
A list with all the different bits of information.  Probably the main bit you care about is \code{ibs$ibs}
}
\description{
Calculates identity by similarity for a series of BAM files. Looks for locations that support the ALT allele in the target regions in a reference sample.  Then counts at those locations in all BAMs to see if the genotypes match.  Intended use is with a \code{regions} parameter containing exons.
}
\details{
If not given, the SNPs are calculated from a reference sample, where heterozygous SNPs are identified.  This only really works well when you have exome or genome sequencing data, otherwise you are better off providing a pre-calculated set of SNPs.
}
