% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseSNPsFromCN.R
\name{phaseSNPsFromCN}
\alias{phaseSNPsFromCN}
\title{Phases heterozygous SNPs in a patient using CN segments}
\usage{
phaseSNPsFromCN(
  hSNPs,
  segs,
  refGenome,
  tBAM,
  outPath = NULL,
  FDR = 0.05,
  minPhasable = 2 * FDR,
  errRate = 0.01,
  useEM = TRUE,
  alleleCounterParams = list(f = 3, F = 3852, m = 20, q = 35),
  nParallel = 1,
  plotMixtures = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{hSNPs}{The heterozygous SNP locations.  See \code{\link{findHetSNPs}}.}

\item{segs}{A GRanges object giving the genomic coordinates where copy number changes that produce allelic imbalance are present.}

\item{refGenome}{Path to reference genome used to map BAM.}

\item{tBAM}{BAM file containing tumour genotype.}

\item{outPath}{File name used to storing genotyping calculations.  If left at NULL, results are not saved and cannot be reloaded in subsequent runs.}

\item{FDR}{False discover rate to accept.}

\item{minPhasable}{If the fraction of phasable SNPs in a segment is less than this, flag the segment as likely improperly called.}

\item{errRate}{Assumed error rate for sanity check that location is heterozygous.  This check is disabled if \code{errRate=0}}

\item{useEM}{Use the expectation-maximisation method to phase SNPs.  See details.}

\item{alleleCounterParams}{Parameters to be used when running allele counter.}

\item{nParallel}{How many threads to use.}

\item{plotMixtures}{If TRUE and using EM mode produces one plot for each CN segment showing the mixture solutions.}

\item{verbose}{Be verbose?  FALSE/0 for not at all, TRUE or 1 for a bit, 2 for very.}

\item{...}{Does nothing.}
}
\value{
\code{hSNPs} but with extra columns indicating results of phasing.
}
\description{
Given heterozygous SNPs in an individual, as called using \code{\link{findHetSNPs}}, and a series of segments where a copy number change has occurred, checks genotype in Tumour BAM files at these locations.  This information is used to identify which allele is major and minor at each SNP.  The major allele is arbitrarily labelled as "maternal" for consistency with downstream functions.
}
\details{
The phasing of SNPs relies on the major allele having a detectable and statistically significant deviation from the other allele.  As such, the larger the allelic imbalance, the more SNPs will be phased.  Note that for tumour samples, the allelic imbalance is determined by the nature of the copy number change and the amount of normal contamination in the tumour biopsy.  For subtle shifts (e.g. 3/2 CN ratio, 80% contamination), it may not be possible to phase many SNPs and those SNPs that are phased may not be particularly reliable.

If \code{useEM=FALSE}, a binomial test to exclude a null hypothesis of allele ratio 0.5 is done and any SNP with false discover rate less than \code{FDR} are phased.  This is almost, but not exactly, equivalent to the default approach, which is to use an EM algorithm to model the SNPs as a mixture of two populations with binomial likelihood.  In this mode, (the default, \code{useEM=TRUE}), the allelic ratio is calculated within each segment and then used to estimate the probability of each SNP belonging to each allele.  This allows a check that the segment is consistent with a CN change being present in the tumour BAM file.  In this mode \code{FDR} is interpreted as how far the allele probability needs to be from 1 to phase a SNP.  For example, \code{FDR=0.01} implies that SNPs with major allele probabilities \code{0.99} and greater or \code{0.01} and lower will be used.  In EM mode the \code{stateProbs} column in the output represents the major allele probability, while when \code{useEM=FALSE} it represents the binomial test p-value.

Note that both BAMs (the one used to generate \code{hSNPs} and \code{tBAM}) must have been mapped to the same reference genome, \code{refGenome}.

Segments are sanity checked for the presence of an allelic imbalance.  If fewer than \code{minPhasable} SNPs can phased, the entire segment is flagged is likely to have failed.
}
