% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCell.R
\name{plotCell}
\alias{plotCell}
\title{Plots the evidence for one cell in detail.}
\usage{
plotCell(
  gCnts,
  cellID,
  regions = gCnts@metadata$segs,
  geneIDs = gCnts$geneName,
  orderBy = "pos",
  spaceBy = NULL,
  pointShape = getShapes(ifelse(gCnts$altIsMum, "matIsALT", "matIsREF")),
  pointColour = getColours(gCnts$regionType),
  plotState = NULL,
  plotGeneNames = FALSE,
  plotGeneLines = FALSE,
  plotLegends = TRUE,
  ncol = NULL,
  nrow = NULL
)
}
\arguments{
\item{gCnts}{Allele specific counts at individual loci in single cell data.  Produced by \code{\link{getAllelicExpression}}.}

\item{cellID}{The ID of the cell to plot evidence for.}

\item{regions}{GRanges regions to plot independently.  If NULL, plot everything split by chromosome.  Can also be a vecetor of chromosomes to plot.}

\item{geneIDs}{List or vector of same length as \code{gCnts}.}

\item{orderBy}{Used to order SNPs in plot.  Anything in \code{gCnts} is valid and start with '-' to reverse order.  'pos' is genomic position.}

\item{spaceBy}{How to space things.  Anything in \code{gCnts} is valid.  'pos' is genomic position.  If NULL, space evenly.  Will space on the value of \code{spaceBy} unless \code{spaceBy} starts with a '-' in which case space on the difference.}

\item{pointShape}{Vector of same length as \code{gCnts} giving the pch value to use for each point.}

\item{pointColour}{Vector of same length as \code{gCnts} giving the colour to use for each point.}

\item{plotState}{The copy number state to plot.  Requires that \code{\link{calcLikelihoods}} have been run on \code{gCnts}.  If NULL, state not plotted.}

\item{plotGeneNames}{Include gene names below.}

\item{plotGeneLines}{Put separating lines at the boundary of each gene?}

\item{plotLegends}{Show legends?}

\item{ncol}{Number of columns to use in array of chromosomes.}

\item{nrow}{Number of columns to use in array of chromosomes.}
}
\value{
A lovely lovely plot.
}
\description{
Plots summary of evidence for an individual cell.  SNPs are ordered such that those that provide the most weight are on the right.  Shape shows if the SNP is in an intron,exon, or intergenic.  The red/black line shows the tumour/normal BAF for each chrommosome.  The dots and error bars show the estimate and confidence interval for the BAF of this SNP aggregating across cells marked as normal.  The dotted line shows the cumulative BAF aggregating across SNPs from the left.
}
