% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoverageAndBAF.R
\name{plotCoverageAndBAF}
\alias{plotCoverageAndBAF}
\title{Plot coverage and BAF}
\usage{
plotCoverageAndBAF(
  hSNPs,
  useLogR = TRUE,
  linesBAF = c(),
  alpha = max(0.002, min(1, 1e+05/length(hSNPs))),
  minCoverage = 10,
  plotASCAT = !is.null(hSNPs@metadata$ascat.output),
  useCorrectedLogR = TRUE,
  chrsToPlot = c(1:22, "X")
)
}
\arguments{
\item{hSNPs}{Output of \code{\link{generateCoverageAndBAF}}}

\item{useLogR}{Should logR be plotted instead of raw coverage?}

\item{linesBAF}{Draw horizontal lines at these BAFs.}

\item{alpha}{Transparency for points used in plotting.}

\item{minCoverage}{Don't plot points with coverage less than this.}

\item{plotASCAT}{Add a third panel showing the ASCAT results.}

\item{useCorrectedLogR}{If GC corrected logR values are available (from ASCAT) and \code{useLogR=TRUE} plot corrected values.}

\item{chrsToPlot}{Which chromosomes to plot.}
}
\value{
Nothing, but makes a pretty plot.
}
\description{
Takes a \code{GRanges} object for which coverage and BAF have been previously calculated using \code{\link{generateCoverageAndBAF}} and plots the coverage, BAF, and ASCAT output (if calculated).
}
