% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPosteriorHeatmap.R
\name{plotPosteriorHeatmap}
\alias{plotPosteriorHeatmap}
\title{A region by cell heatmap of posterior probabilities}
\usage{
plotPosteriorHeatmap(
  pp,
  plotType = c("postProb", "nLL"),
  refState = NULL,
  plotStates = NULL,
  useRemoveGroup = FALSE,
  zLims = c(-5, 5),
  colPal = ifelse(plotType == "postProb", "Greens", "BrBG"),
  groupCellsBy = pp$clusterID,
  ...
)
}
\arguments{
\item{pp}{Posterior probabilities produces by \code{\link{calcStateProbs}}.}

\item{plotType}{Either 'nLL' to plot the negative log-likelihoods or 'postProb' for posterior probabilities.}

\item{refState}{If type 'nLL', plot differences in log-likelihood relative to this state.  Ignored otherwise or if NULL.}

\item{plotStates}{Plot posterior probability for these states.  NULL uses all states if >2 states or first if 2.}

\item{useRemoveGroup}{Should stats be extracted from the group where N have been left out?}

\item{zLims}{Truncate plot values to this range.  If \code{plotType} is 'postProb', transformed to on the interval (0,1) with an inverse logit transformation.}

\item{colPal}{Colour scheme to use.  Passed to \code{\link[RColorBrewer]{brewer.pal}}.}

\item{groupCellsBy}{A vector of the same length as \code{pp} indicating how to group cells.}

\item{...}{Extra parameters passed to \code{\link[ComplexHeatmap]{Heatmap}}.}
}
\value{
A plot, obvs.
}
\description{
Takes the output of \code{\link{calcStateProbs}} and produces a heatmap showing the posterior probabilities for each cell in each region.  The rows and columns of the heatmap are constructed from \code{factor(pp$cellID)} and \code{factor(pp$regionID)}.  The order can be changed, and extras such as row/column splits provided, by pre-setting these columns of \code{pp} to factors and supplying entries based on their levels.
}
\details{
This is essentialy just a smart wrapper around the \code{\link[ComplexHeatmap]{Heatmap}} function, so you can pass any of the same arguments to this function.
}
