% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRawData.R
\name{plotRawData}
\alias{plotRawData}
\title{Plots a high level summary of raw data}
\usage{
plotRawData(
  gCnts,
  clusterID = gCnts$clusterID,
  segs = gCnts@metadata$segs,
  tgtMAF = segs$tumFrac,
  returnData = FALSE,
  showPlot = returnData
)
}
\arguments{
\item{gCnts}{GRanges object with single cell level count data.  Usually output of \code{\link{getAllelicExpression}}.}

\item{clusterID}{Vector of the same length as \code{gCnts} indicating how to group cells into clusters.  If NULL, all cells treated as one cluster.}

\item{segs}{Genomic segments within which data will be summarised.  If NULL, summarised by chromosome.}

\item{tgtMAF}{The target maternal allele fraction for each segment.  Must be same length as \code{segs} or NULL, in which case it will default to 0.5 everywhere.}

\item{returnData}{Return data as well as plot object.}

\item{showPlot}{Don't just return the plot object, explicitly plot it.}
}
\value{
Either a ggplot object or if \code{returnData=TRUE} a list containing the cell and cluster summary tables and the ggplot object for the plot.
}
\description{
Summarises data at the cell (and cluster if available) level by either chromosome or segment specified.  The default plot shows how each cluster of cells behave in each segment.  Each point represents a cell, the boxplot shows the distribution, and the big red points show the value treating each cluster as one cell.  The colour scheme shows low/medium/high coverage for each cell/segment combination.
}
\details{
If \code{returnData=TRUE}, this function will return two tables that give the summary statistics for each cell/segment combination (\code{cellCnts}) and cluster/segment combination (\code{clustCnts}).
}
