% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSplit.R
\name{plotSplit}
\alias{plotSplit}
\title{Plot normalised counts split by some variable}
\usage{
plotSplit(
  gCnts,
  assays = c("matCount", "patCount"),
  assayCols = c("darkred", "purple"),
  transposeMatrix = FALSE,
  orderBy = "counts",
  countMax = max,
  countTrans = log10,
  countTransLab = "(log10)",
  xlab = ifelse(transposeMatrix, "cell", "region"),
  plotNames = NULL,
  mar = c(5, 4, 4, 5),
  lwdInside = 0.1,
  lwdBox = 0.5,
  showLegends = TRUE
)
}
\arguments{
\item{gCnts}{Allele specific counts at individual loci in single cell data.  Produced by \code{\link{getAllelicExpression}}.}

\item{assays}{Columns in \code{gCnts} to aggregate over.}

\item{assayCols}{Colours for assays.  Must be the same length as \code{assayCols}.}

\item{transposeMatrix}{Should count matricies built from assays be transposed?  This swaps the meaning of rows and columns.}

\item{orderBy}{How to order columns.  Should be either the name of an assay or 'counts' for total counts.}

\item{countMax}{Function to apply to vector of transformed counts to get the upper limit for plotting.}

\item{countTrans}{What transformation should be applied to the counts?}

\item{countTransLab}{What label to give the count legend?}

\item{xlab}{Label for x-axis.}

\item{plotNames}{Should we plot the names of the columns?  If NULL, plots if fewer than 20 entries.}

\item{mar}{Margins for plot.}

\item{lwdInside}{The width for lines separating columns.}

\item{lwdBox}{The width for lines separating parts of columns.}

\item{showLegends}{Plot legends to right of plot?}
}
\value{
A plot, obs.
}
\description{
This function will plot one bar plot per grouping, which is normalised to sum to 1.  Each bar is split into groups of different colours and further sub-divisions indicated by black boxes.  The default is to show some genomic grouping (typically het SNPs or genes), colour the bar by maternal/paternal counts, and indicate contributions from individual cells with boxes.
}
\details{
This seemed like a good idea at the time.  I'm sure there was some point to this, but I can't remember what it was.
}
